Found and fix single city bug
Had a bug where if I entered a city for search where the only result
is one--say, 'Boca Raton' or 'West Palm Beach', the app would
have a problem. Problem was that there was a creation of
elements that were undefined. This is because the code
assumed there would be three and only three responses, never
less than that. Fixed that by putting logic into the render function
of view obj that adjusts based on length of array.